make usernames work IT WORKKSSSS!
client will prompt user for a username, send it (along with a prefix tag thing)
to the server. The server will save it with your client address. When you send
a message to the server, if it has your address in the library, it will display
your username instead of address.

I've essentially made a rudimentary (currently password-less) login system, and
it works with multiple clients.