impliment rsa auth and fernet encryption We are now able to have a client authenticate by sending an rsa public key to the server. The server sends a symmetric key back to the client, encrypted using the rsa key.
The client then decrypts it with their rsa private key. They then are able to communicate back
and forth with the server using the symmetric key.

The server can do this process with as many clients that join, and it will use a unique
symmetric key for each client-session.

We need to next work on disconnecting users who have not sent a message to the server
in some length of time (presumably, they've already disconnected on their end, but
they did so by closing the terminal, so it never sent the disconnect message).
They might've also just been idling, which could lag the server if it became
a large-scale platform.